Move config view model into common module
