+++ wallet/src/main/java/ch/admin/bag/covidcertificate/wallet/transfercode/TransferCodeHowToFragment.kt
@@ -16,12 +16,13 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import ch.admin.bag.covidcertificate.common.config.ConfigViewModel
 import ch.admin.bag.covidcertificate.common.config.FaqModel
 import ch.admin.bag.covidcertificate.common.faq.FaqAdapter
 import ch.admin.bag.covidcertificate.common.faq.model.Faq
 import ch.admin.bag.covidcertificate.common.faq.model.Header
 import ch.admin.bag.covidcertificate.common.faq.model.Question
-import ch.admin.bag.covidcertificate.wallet.ConfigViewModel
+import ch.admin.bag.covidcertificate.wallet.BuildConfig
 import ch.admin.bag.covidcertificate.wallet.R
 import ch.admin.bag.covidcertificate.wallet.databinding.FragmentTransferCodeHowtoBinding
 
@@ -53,7 +54,7 @@ class TransferCodeHowToFragment : Fragment(R.layout.fragment_transfer_code_howto
 				showTransferCodeFaqItems(it)
 			}
 		}
-		configViewModel.loadConfig()
+		configViewModel.loadConfig(BuildConfig.BASE_URL, BuildConfig.VERSION_NAME, BuildConfig.BUILD_TIME.toString())
 	}
 
 	override fun onDestroyView() {
